+ register_generate_api_routers
